Replace commander with yargs
