+++ package.json
@@ -21,7 +21,7 @@
   "homepage": "https://github.com/financial-times-sandbox/manage-github-apps#readme",
   "dependencies": {
     "@octokit/rest": "^15.13.1",
-    "commander": "^2.19.0"
+    "yargs": "^12.0.2"
   },
   "devDependencies": {
     "eslint": "^5.7.0",
